Move ValidationSettings into a separate file and disable validation by default
It was found that the validation package can cause annoyances when introducing it to an existing project or for users who do not want to be affected by error messages. For example, during the process of upgrading a project to use the Optional attribute, users may not be able to enter play mode when validation is enabled, although the missing references are not necessarily real errors. So, by default, the validation should be disabled for new users and then be opt-in for those who wish to use it and know how to configure it, etc.